feat: rename quickjob, quickcronjob and quickdeployment
