+++ src/lib/components/kind-viewer/kind-viewer-actions/schedule/actions.svelte
@@ -1,15 +1,13 @@
 <script lang="ts">
-	import EllipsisIcon from '@lucide/svelte/icons/ellipsis';
+	import Ellipsis from '@lucide/svelte/icons/ellipsis';
 
 	import Delete from '$lib/components/kind-viewer/kind-viewer-actions/default/delete.svelte';
 	import Describe from '$lib/components/kind-viewer/kind-viewer-actions/default/describe.svelte';
-	import Edit from '$lib/components/kind-viewer/kind-viewer-actions/default/edit.svelte';
 	import View from '$lib/components/kind-viewer/kind-viewer-actions/default/view.svelte';
-	import UpdateCronjob from '$lib/components/kind-viewer/kind-viewer-actions/kro/update-cronjob.svelte';
-	import UpdateDeployment from '$lib/components/kind-viewer/kind-viewer-actions/kro/update-deployment.svelte';
-	import UpdateJob from '$lib/components/kind-viewer/kind-viewer-actions/kro/update-job.svelte';
 	import Button from '$lib/components/ui/button/button.svelte';
 	import * as DropdownMenu from '$lib/components/ui/dropdown-menu/index.js';
+
+	import Edit from './update.svelte';
 
 	let {
 		schema,
@@ -38,8 +36,8 @@
 	<DropdownMenu.Trigger>
 		{#snippet child({ props })}
 			<div class="flex justify-end">
-				<Button size="icon" variant="ghost" class="shadow-none" aria-label="Edit item" {...props}>
-					<EllipsisIcon size={16} aria-hidden="true" />
+				<Button size="icon" variant="ghost" class="shadow-none" aria-label="Actions" {...props}>
+					<Ellipsis size={16} aria-hidden="true" />
 				</Button>
 			</div>
 		{/snippet}
@@ -58,17 +56,7 @@
 					e.preventDefault();
 				}}
 			>
-				<Describe
-					{cluster}
-					{namespace}
-					{group}
-					{version}
-					{resource}
-					{object}
-					onOpenChangeComplete={() => {
-						actionsOpen = false;
-					}}
-				/>
+				<Describe {cluster} {namespace} {group} {version} {resource} {object} />
 			</DropdownMenu.Item>
 			<DropdownMenu.Item
 				onSelect={(e) => {
@@ -88,55 +76,6 @@
 						actionsOpen = false;
 					}}
 				/>
-			</DropdownMenu.Item>
-			<DropdownMenu.Item
-				onSelect={(e) => {
-					e.preventDefault();
-				}}
-			>
-				{#if kind === 'QuickDeployment' || kind === 'quickDeployment'}
-					<UpdateDeployment
-						{cluster}
-						{namespace}
-						{group}
-						{version}
-						{kind}
-						{resource}
-						{schema}
-						{object}
-						onOpenChangeComplete={() => {
-							actionsOpen = false;
-						}}
-					/>
-				{:else if kind === 'QuickJob' || kind === 'quickJob'}
-					<UpdateJob
-						{cluster}
-						{namespace}
-						{group}
-						{version}
-						{kind}
-						{resource}
-						{schema}
-						{object}
-						onOpenChangeComplete={() => {
-							actionsOpen = false;
-						}}
-					/>
-				{:else if kind === 'QuickCronJob' || kind === 'quickCronjob'}
-					<UpdateCronjob
-						{cluster}
-						{namespace}
-						{group}
-						{version}
-						{kind}
-						{resource}
-						{schema}
-						{object}
-						onOpenChangeComplete={() => {
-							actionsOpen = false;
-						}}
-					/>
-				{/if}
 			</DropdownMenu.Item>
 			<DropdownMenu.Item
 				onSelect={(e) => {
