+++ src/lib/components/kind-viewer/kind-viewer-actions/task/create.svelte
@@ -1,9 +1,11 @@
 <script lang="ts">
 	import { ConnectError, createClient, type Transport } from '@connectrpc/connect';
-	import FormIcon from '@lucide/svelte/icons/form-input';
+	import Plus from '@lucide/svelte/icons/plus';
 	import { ResourceService } from '@otterscale/api/resource/v1';
 	import type { FormValue, Schema, UiSchemaRoot } from '@sjsf/form';
 	import { SubmitButton } from '@sjsf/form';
+	import Ajv from 'ajv';
+	import { load } from 'js-yaml';
 	import lodash from 'lodash';
 	import { mode as themeMode } from 'mode-watcher';
 	import { getContext } from 'svelte';
@@ -26,7 +28,6 @@
 		kind,
 		resource,
 		schema: jsonSchema,
-		object,
 		onOpenChangeComplete
 	}: {
 		cluster: string;
@@ -35,36 +36,24 @@
 		version: string;
 		kind: string;
 		resource: string;
-		schema: any;
-		object: any;
+		schema?: any;
 		onOpenChangeComplete: () => void;
 	} = $props();
 
 	const transport: Transport = getContext('transport');
 	const resourceClient = createClient(ResourceService, transport);
 
-	const systemFields = [
-		'clusterName',
-		'creationTimestamp',
-		'deletionGracePeriodSeconds',
-		'deletionTimestamp',
-		'finalizers',
-		'generateName',
-		'generation',
-		'initializers',
-		'managedFields',
-		'ownerReferences',
-		'resourceVersion',
-		'relationships',
-		'selfLink',
-		'state',
-		'uid'
-	];
+	const jsonSchemaValidator = new Ajv({
+		allErrors: true,
+		strict: false
+	});
+	const validate = jsonSchemaValidator.compile(jsonSchema);
 
+	// Container for Data — flat structure matching quickJob RGD schema
 	let values: any = $state({
 		apiVersion: group ? `${group}/${version}` : version,
 		kind,
-		metadata: object.metadata,
+		metadata: {},
 		spec: {}
 	});
 	let settingsValues: any = $state({});
@@ -75,18 +64,10 @@
 		values.spec = { ...settingsValues, ...specValues, ...resourceValues };
 	});
 
-	let value = $state('');
-	$effect(() => {
-		const filtered = lodash.cloneDeep(values);
-		if (filtered.metadata) {
-			for (const field of systemFields) {
-				delete filtered.metadata[field];
-			}
-		}
-		delete filtered.status;
-		value = stringify(filtered);
-	});
-	const steps = Array.from({ length: 4 }, (_, index) => String(index + 1));
+	let value = $derived(stringify(values));
+
+	// Steps: 1=Metadata, 2=Setting, 3=Container, 4=Resources, 5=YAML Preview
+	const steps = Array.from({ length: 5 }, (_, index) => String(index + 1));
 	const [firstStep] = steps;
 	let currentStep = $state(firstStep);
 	const currentIndex = $derived(steps.indexOf(currentStep));
@@ -100,6 +81,7 @@
 		currentStep = firstStep;
 	}
 
+	// Flag for Dialog
 	let open = $state(false);
 	let isSubmitting = $state(false);
 </script>
@@ -115,14 +97,9 @@
 >
 	<Dialog.Trigger>
 		{#snippet child({ props })}
-			<Item.Root {...props} class="w-full p-0 text-xs" size="sm">
-				<Item.Media>
-					<FormIcon />
-				</Item.Media>
-				<Item.Content>
-					<Item.Title>Update</Item.Title>
-				</Item.Content>
-			</Item.Root>
+			<Button {...props} variant="outline" size="icon">
+				<Plus />
+			</Button>
 		{/snippet}
 	</Dialog.Trigger>
 	<Dialog.Content
@@ -137,53 +114,92 @@
 			</Item.Content>
 		</Item.Root>
 		<Tabs.Root value={currentStep}>
-			<!-- Step 1: Setting -->
+			<!-- Step 1: Metadata -->
 			<Tabs.Content value={steps[0]}>
 				<Form
 					schema={{
+						...(lodash.omit(lodash.get(jsonSchema, 'properties.metadata'), 'properties') as Record<
+							string,
+							unknown
+						>),
+						title: 'Metadata',
+						properties: {
+							name: {
+								...lodash.get(jsonSchema, 'properties.metadata.properties.name'),
+								title: 'Name'
+							},
+							namespace: {
+								...lodash.get(jsonSchema, 'properties.metadata.properties.namespace'),
+								title: 'Namespace',
+								readOnly: true
+							}
+						}
+					} as Schema}
+					uiSchema={{
+						'ui:options': {
+							translations: {
+								submit: 'Next'
+							}
+						}
+					} as UiSchemaRoot}
+					initialValue={{ namespace: namespace } as FormValue}
+					handleSubmit={{
+						posthook: () => {
+							handleNext();
+						}
+					}}
+					bind:values={values['metadata']}
+				>
+					{#snippet actions()}
+						<div class="flex w-full items-center justify-between gap-3">
+							<Button
+								onclick={() => {
+									handlePrevious();
+								}}
+								disabled={currentIndex === 0}
+							>
+								Previous
+							</Button>
+							<SubmitButton />
+						</div>
+					{/snippet}
+				</Form>
+			</Tabs.Content>
+
+			<!-- Step 2: Setting -->
+			<Tabs.Content value={steps[1]}>
+				<Form
+					schema={{
 						title: 'Setting',
-						type: 'object',
+						...lodash.omit(lodash.get(jsonSchema, 'properties.spec'), 'properties'),
 						properties: {
 							completions: {
-								type: 'integer',
-								title: 'Completions',
-								default: 1,
-								minimum: 0
+								...lodash.get(jsonSchema, 'properties.spec.properties.completions'),
+								title: 'Completions'
 							},
 							parallelism: {
-								type: 'integer',
-								title: 'Parallelism',
-								default: 1,
-								minimum: 0
+								...lodash.get(jsonSchema, 'properties.spec.properties.parallelism'),
+								title: 'Parallelism'
 							},
 							backoffLimit: {
-								type: 'integer',
-								title: 'Backoff Limit',
-								default: 6,
-								minimum: 0
+								...lodash.get(jsonSchema, 'properties.spec.properties.backoffLimit'),
+								title: 'Backoff Limit'
 							},
 							activeDeadlineSeconds: {
-								type: 'integer',
-								title: 'Active Deadline Seconds',
-								default: 0,
-								minimum: 0
+								...lodash.get(jsonSchema, 'properties.spec.properties.activeDeadlineSeconds'),
+								title: 'Active Deadline Seconds'
 							},
 							ttlSecondsAfterFinished: {
-								type: 'integer',
-								title: 'TTL Seconds After Finished',
-								default: 0,
-								minimum: 0
+								...lodash.get(jsonSchema, 'properties.spec.properties.ttlSecondsAfterFinished'),
+								title: 'TTL Seconds After Finished'
 							},
 							suspend: {
-								type: 'boolean',
-								title: 'Suspend',
-								default: false
+								...lodash.get(jsonSchema, 'properties.spec.properties.suspend'),
+								title: 'Suspend'
 							},
 							restartPolicy: {
-								type: 'string',
-								title: 'Restart Policy',
-								default: 'Never',
-								enum: ['Never', 'OnFailure']
+								...lodash.get(jsonSchema, 'properties.spec.properties.restartPolicy'),
+								title: 'Restart Policy'
 							}
 						}
 					} as Schema}
@@ -215,13 +231,13 @@
 						}
 					} as UiSchemaRoot}
 					initialValue={{
-						completions: object.spec?.completions ?? 1,
-						parallelism: object.spec?.parallelism ?? 1,
-						backoffLimit: object.spec?.backoffLimit ?? 6,
-						activeDeadlineSeconds: object.spec?.activeDeadlineSeconds ?? 0,
-						ttlSecondsAfterFinished: object.spec?.ttlSecondsAfterFinished ?? 0,
-						suspend: object.spec?.suspend ?? false,
-						restartPolicy: object.spec?.restartPolicy ?? 'Never'
+						completions: 1,
+						parallelism: 1,
+						backoffLimit: 6,
+						activeDeadlineSeconds: 3600,
+						ttlSecondsAfterFinished: 86400,
+						suspend: false,
+						restartPolicy: 'Never'
 					} as FormValue}
 					handleSubmit={{
 						posthook: () => {
@@ -236,7 +252,6 @@
 								onclick={() => {
 									handlePrevious();
 								}}
-								disabled={currentIndex === 0}
 							>
 								Previous
 							</Button>
@@ -246,38 +261,35 @@
 				</Form>
 			</Tabs.Content>
 
-			<!-- Step 2: Container -->
-			<Tabs.Content value={steps[1]}>
+			<!-- Step 3: Container (name, image, command, args, containerPort) -->
+			<Tabs.Content value={steps[2]}>
 				<Form
 					schema={{
 						title: 'Container',
 						type: 'object',
-						required: ['name', 'image'],
+						required: lodash
+							.get(jsonSchema, 'properties.spec.required', [])
+							.filter((f: string) => ['name', 'image'].includes(f)),
 						properties: {
 							name: {
-								type: 'string',
-								title: 'Name',
-								description: 'Name of the workload'
+								...lodash.get(jsonSchema, 'properties.spec.properties.name'),
+								title: 'Name'
 							},
 							image: {
-								type: 'string',
-								title: 'Image',
-								description: 'Container image to use'
+								...lodash.get(jsonSchema, 'properties.spec.properties.image'),
+								title: 'Image'
 							},
 							command: {
-								type: 'array',
-								title: 'Command',
-								items: { type: 'string' }
+								...lodash.get(jsonSchema, 'properties.spec.properties.command'),
+								title: 'Command'
 							},
 							args: {
-								type: 'array',
-								title: 'Arguments',
-								items: { type: 'string' }
+								...lodash.get(jsonSchema, 'properties.spec.properties.args'),
+								title: 'Arguments'
 							},
 							containerPort: {
-								type: 'integer',
-								title: 'Container Port',
-								default: 8080
+								...lodash.get(jsonSchema, 'properties.spec.properties.containerPort'),
+								title: 'Container Port'
 							}
 						}
 					} as Schema}
@@ -299,11 +311,9 @@
 						}
 					} as UiSchemaRoot}
 					initialValue={{
-						name: object.spec?.name ?? null,
-						image: object.spec?.image ?? null,
-						command: object.spec?.command ?? undefined,
-						args: object.spec?.args ?? undefined,
-						containerPort: object.spec?.containerPort ?? 8080
+						name: values.metadata?.name ?? null,
+						image: null,
+						containerPort: 8080
 					} as FormValue}
 					handleSubmit={{
 						posthook: () => {
@@ -318,7 +328,6 @@
 								onclick={() => {
 									handlePrevious();
 								}}
-								disabled={currentIndex === 0}
 							>
 								Previous
 							</Button>
@@ -328,42 +337,36 @@
 				</Form>
 			</Tabs.Content>
 
-			<!-- Step 3: Resources -->
-			<Tabs.Content value={steps[2]}>
+			<!-- Step 4: Resources -->
+			<Tabs.Content value={steps[3]}>
 				<Form
 					schema={{
 						title: 'Resources',
 						type: 'object',
 						properties: {
 							resourcesRequestsCpu: {
-								type: 'string',
-								title: 'CPU Request',
-								default: '100m'
+								...lodash.get(jsonSchema, 'properties.spec.properties.resourcesRequestsCpu'),
+								title: 'CPU Request'
 							},
 							resourcesRequestsMemory: {
-								type: 'string',
-								title: 'Memory Request',
-								default: '128Mi'
+								...lodash.get(jsonSchema, 'properties.spec.properties.resourcesRequestsMemory'),
+								title: 'Memory Request'
 							},
 							resourcesLimitsCpu: {
-								type: 'string',
-								title: 'CPU Limit',
-								default: '500m'
+								...lodash.get(jsonSchema, 'properties.spec.properties.resourcesLimitsCpu'),
+								title: 'CPU Limit'
 							},
 							resourcesLimitsMemory: {
-								type: 'string',
-								title: 'Memory Limit',
-								default: '512Mi'
+								...lodash.get(jsonSchema, 'properties.spec.properties.resourcesLimitsMemory'),
+								title: 'Memory Limit'
 							},
 							resourcesGpu: {
-								type: 'string',
-								title: 'GPU',
-								default: '0'
+								...lodash.get(jsonSchema, 'properties.spec.properties.resourcesGpu'),
+								title: 'GPU'
 							},
 							resourcesGpumem: {
-								type: 'string',
-								title: 'GPU Memory',
-								default: '0'
+								...lodash.get(jsonSchema, 'properties.spec.properties.resourcesGpumem'),
+								title: 'GPU Memory'
 							}
 						}
 					} as Schema}
@@ -380,12 +383,12 @@
 						}
 					} as UiSchemaRoot}
 					initialValue={{
-						resourcesRequestsCpu: object.spec?.resourcesRequestsCpu ?? '100m',
-						resourcesRequestsMemory: object.spec?.resourcesRequestsMemory ?? '128Mi',
-						resourcesLimitsCpu: object.spec?.resourcesLimitsCpu ?? '500m',
-						resourcesLimitsMemory: object.spec?.resourcesLimitsMemory ?? '512Mi',
-						resourcesGpu: object.spec?.resourcesGpu ?? '0',
-						resourcesGpumem: object.spec?.resourcesGpumem ?? '0'
+						resourcesRequestsCpu: '100m',
+						resourcesRequestsMemory: '128Mi',
+						resourcesLimitsCpu: '500m',
+						resourcesLimitsMemory: '512Mi',
+						resourcesGpu: '0',
+						resourcesGpumem: '0'
 					} as FormValue}
 					handleSubmit={{
 						posthook: () => {
@@ -409,8 +412,8 @@
 				</Form>
 			</Tabs.Content>
 
-			<!-- Step 4: YAML Preview & Submit -->
-			<Tabs.Content value={steps[3]} class="min-h-[77vh]">
+			<!-- Step 5: YAML Preview & Submit -->
+			<Tabs.Content value={steps[4]} class="min-h-[77vh]">
 				<div class="flex h-full flex-col gap-3">
 					<Monaco
 						options={{
@@ -432,32 +435,48 @@
 
 							isSubmitting = true;
 
-							const name = object.metadata?.name;
-							const manifest = new TextEncoder().encode(value);
+							let parsed: any;
+							try {
+								parsed = load(value);
+							} catch {
+								toast.error('Invalid YAML. Please check the content.');
+								isSubmitting = false;
+								return;
+							}
+
+							const isValid = validate(parsed);
+
+							if (!isValid) {
+								console.error('Validation errors:', validate.errors);
+								toast.error('Validation failed. Please check the YAML.');
+								isSubmitting = false;
+								return;
+							}
+
+							const name = lodash.get(parsed, 'metadata.name');
 
 							toast.promise(
 								async () => {
-									await resourceClient.apply({
+									const manifest = new TextEncoder().encode(value);
+
+									await resourceClient.create({
 										cluster,
 										namespace,
 										group,
 										version,
 										resource,
-										name,
-										manifest,
-										fieldManager: 'otterscale-web-ui',
-										force: true
+										manifest
 									});
 								},
 								{
-									loading: `Updating ${kind} ${name}...`,
+									loading: `Creating ${kind} ${name}...`,
 									success: () => {
 										open = false;
-										return `Successfully updated ${kind} ${name}`;
+										return `Successfully created ${kind} ${name}`;
 									},
 									error: (error) => {
-										console.error(`Failed to update ${kind} ${name}:`, error);
-										return `Failed to update ${kind} ${name}: ${(error as ConnectError).message}`;
+										console.error(`Failed to create ${kind} ${name}:`, error);
+										return `Failed to create ${kind} ${name}: ${(error as ConnectError).message}`;
 									},
 									finally() {
 										isSubmitting = false;
@@ -466,7 +485,7 @@
 							);
 						}}
 					>
-						Update
+						Create
 					</Button>
 				</div>
 			</Tabs.Content>
